Custom element instance should be reused for injected by @Id element
Share custom element instance created by a template parser with @Id
injected field if needed.

Fixes #1812
