+++ flow-tests/test-root-context/src/main/webapp/com/vaadin/flow/uitest/ui/template/InjectedChild.html
@@ -1,0 +1,29 @@
+<!--
+  ~ Copyright 2000-2017 Vaadin Ltd.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License"); you may not
+  ~ use this file except in compliance with the License. You may obtain a copy of
+  ~ the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+  ~ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+  ~ License for the specific language governing permissions and limitations under
+  ~ the License.
+  -->
+<link rel="import" href="/bower_components/polymer/polymer.html">
+
+<dom-module id="injected-child">
+    <template>
+        <div id="text">[[text]]</div>
+    </template>
+    <script>
+        class InjectedChild extends Polymer.Element {
+            static get is() { return 'injected-child' }
+           
+        }
+        customElements.define(InjectedChild.is, InjectedChild);
+    </script>
+</dom-module>
