+++ flow-tests/test-root-context/src/main/webapp/com/vaadin/flow/uitest/ui/template/ParentTemplateInjectChild.html
@@ -1,0 +1,33 @@
+<!--
+  ~ Copyright 2000-2017 Vaadin Ltd.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License"); you may not
+  ~ use this file except in compliance with the License. You may obtain a copy of
+  ~ the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+  ~ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+  ~ License for the specific language governing permissions and limitations under
+  ~ the License.
+  -->
+
+<link rel="import" href="/bower_components/polymer/polymer.html">
+
+<dom-module id="parent-inject-child">
+    <template>
+        <injected-child id='child'></injected-child>
+        <button on-click="updateChild" id="button">Update Child</button>
+    </template>
+
+    <script>
+        class ParentInjectChild extends Polymer.Element {
+            static get is() {
+                return 'parent-inject-child'
+            }
+        }
+        customElements.define(ParentInjectChild.is, ParentInjectChild);
+    </script>
+</dom-module>
